gcc_comp_m1_v4/requirements.txt
# L3 M1.4: Compliance Documentation & Evidence
# Core dependencies for compliance audit trails and evidence collection

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Environment and configuration
python-dotenv==1.0.0

# PostgreSQL database (for audit trail storage)
psycopg2-binary==2.9.9

# AWS S3 (for evidence storage)
boto3==1.34.0

# Jupyter notebook support
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.27.1

# Data handling
pandas==2.1.4

# Logging and monitoring
python-json-logger==2.0.7
